update and deleting authorization done

+++ src/events/events.controller.ts
@@ -1,10 +1,6 @@
-import { Controller, Get, Post, Patch, Delete, Param, Body, HttpCode, ParseIntPipe, ValidationPipe, Logger, NotFoundException, Query, UsePipes, UseGuards } from "@nestjs/common";
+import { Controller, Get, Post, Patch, Delete, Param, Body, HttpCode, ParseIntPipe, ValidationPipe, NotFoundException, Query, UsePipes, UseGuards, ForbiddenException } from "@nestjs/common";
 import { CreateEventDto } from "./input/create-event.dto";
 import { UpdateEventDto } from "./input/udpate-event.dto";
-import { Event } from "./event.entity";
-import { Like, MoreThan, Repository } from "typeorm";
-import { InjectRepository } from "@nestjs/typeorm";
-import { Attendee } from "./attendee.entity";
 import { EventsService } from "./events.service";
 import { ListEvents } from "./input/list.events";
 import { CurrentUser } from "src/auth/current-user.decorator";
@@ -13,14 +9,8 @@ import { AuthGaurdJwt } from "src/auth/auth-gaurd.jwt";
 
 @Controller('/events')
 export class EventsController {
-  private readonly logger = new Logger(EventsController.name);
-
 
   constructor(
-    @InjectRepository(Event)
-    private readonly repository: Repository<Event>,
-    @InjectRepository(Attendee)
-    private readonly attendeerepository: Repository<Attendee>,
     private readonly eventsService: EventsService
   ) { }
 
@@ -39,25 +29,25 @@ export class EventsController {
     return events;
   }
   
-  @Get('/practice')
-  async practice() {
-    return await this.repository.find({
-      select: ['id', 'when'],
-      where: [{
-        id: MoreThan(3),
-        when: MoreThan(new Date('2021-02-12T13:00:00'))
-      }, {
-        description: Like('%meet%')
-      }],
-      take: 2,
-      order: {
-        id: 'DESC'
-      }
-    });
-  }
+  // @Get('/practice')
+  // async practice() {
+  //   return await this.repository.find({
+  //     select: ['id', 'when'],
+  //     where: [{
+  //       id: MoreThan(3),
+  //       when: MoreThan(new Date('2021-02-12T13:00:00'))
+  //     }, {
+  //       description: Like('%meet%')
+  //     }],
+  //     take: 2,
+  //     order: {
+  //       id: 'DESC'
+  //     }
+  //   });
+  // }
   
-  @Get('/practice2')
-  async practice2() {
+  // @Get('/practice2')
+  // async practice2() {
     // return await this.repository.findOne({
     //   where: {
     //     id: 1
@@ -65,22 +55,22 @@ export class EventsController {
     //   relations: ['attendees']
     // });
 
-    const event = await this.repository.findOne({
-      where: {
-        id: 1,
-      },
-      relations: ['attendees']
-    });
-    const attendee = new Attendee();
-    attendee.name = 'Using cascade';
-    // attendee.event = event;
+  //   const event = await this.repository.findOne({
+  //     where: {
+  //       id: 1,
+  //     },
+  //     relations: ['attendees']
+  //   });
+  //   const attendee = new Attendee();
+  //   attendee.name = 'Using cascade';
+  //   // attendee.event = event;
 
-    event.attendees.push(attendee);
+  //   event.attendees.push(attendee);
 
-    // await this.attendeerepository.save(attendee)
-    await this.repository.save(event);
-    return event;
-  }
+  //   // await this.attendeerepository.save(attendee)
+  //   await this.repository.save(event);
+  //   return event;
+  // }
 
   @Get(':id')
   async findOne(@Param('id', ParseIntPipe) id: number) {
@@ -104,26 +94,45 @@ export class EventsController {
   }
 
   @Patch(':id')
-  async update(@Param('id') id, @Body() input: UpdateEventDto) {
-    const event = await this.repository.findOne(id);
+  @UseGuards(AuthGaurdJwt)
+  async update(
+    @Param('id') id,
+    @Body() input: UpdateEventDto,
+    @CurrentUser() user: User
+  ) {
+    const event = await this.eventsService.getEvent(id);
     if (!event) {
       throw new NotFoundException();
     }
-    return await this.repository.save({
-      ...event,
-      ...input,
-      when: input.when ? new Date(input.when) : event.when
-    });
+
+    if(event.organizerID !== user.id) {
+      throw new ForbiddenException(
+        null, "You are not authorized to modify this event!"
+      );
+    }
+    return await this.eventsService.updateEvent(event, input);
   }
 
   @Delete(':id')
+  @UseGuards(AuthGaurdJwt)
   @HttpCode(204)
-  async remove(@Param('id') id) {
-    const result = await this.eventsService.deleteEvent(id)
+  async remove(
+    @Param('id') id,
+    @CurrentUser() user: User,
+  ) {
 
-    if(result?.affected !==1) {
+    const event = await this.eventsService.getEvent(id);
+    if (!event) {
       throw new NotFoundException();
     }
+
+    if(event.organizerID !== user.id) {
+      throw new ForbiddenException(
+        null, "You are not authorized to delete this event!"
+      );
+    }
+
+    await this.eventsService.deleteEvent(id);
     
   }
